fixed db arg load and sqlite pragma setting
